+++ Scripts/sample_query.py
@@ -1,6 +1,5 @@
 import sys
 import os
-
 from pyspark.sql import SparkSession
 
 if __name__ == "__main__":
