+++ protein_sequences/ProtSeq/defproteinsequence.py
@@ -1,0 +1,6 @@
+class ProteinSequence:
+    def __init__(self, accession: str, geneName: str, specie: str, sequence: str):
+        self.accession = accession
+        self.geneName = geneName
+        self.specie = specie
+        self.sequence = sequence
