+++ protein_sequences/ProtSeqIO/fasta_helper.py
